requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==6.0.1
django-allauth==65.14.0
django-debug-toolbar==6.2.0
django-environ==0.12.0
django-multiupload==0.6.1
django-phonenumber-field==8.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.3
environ==1.0
idna==3.11
phonenumbers==9.0.22
phonenumberslite==9.0.22
pillow==12.1.0
psycopg2==2.9.11
pycparser==3.0
PyJWT==2.8.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
transliterate==1.10.2
tzdata==2025.3
urllib3==2.6.3
